163405: fix termintae instance script

+++ scripts/terminate-instance.ps1
@@ -6,7 +6,11 @@ param (
 
     [Parameter(Mandatory=$true)]
     [string]
-    $version
+    $version,
+
+    [Parameter(Mandatory=$false)]
+    [boolean]
+    $DeleteSnapShot=$false
   )
 
 
@@ -23,11 +27,13 @@ Write-Host "Deleting the security group"
 Start-Sleep -Seconds 180
  Remove-EC2SecurityGroup -GroupName "$($version)$($versionText)$($env:BUILD_BUILDNUMBER)" -Force
 
-#Deregister ami and delete snapshot Id
-$ami = "$($env:BUILDIMAGE_AMIID)"
-$snapshot = (Get-EC2Snapshot -owner self | Where-Object {$_.Description -like "*$ami*"}).SnapshotId
-Write-Host "Deregistering the AMI"
-Unregister-EC2Image -ImageId $ami -Force
-Start-Sleep -Seconds 20
-Write-Host "Deleting Snapshot"
-Remove-EC2Snapshot -SnapshotId $snapshot -Force
+if ( $DeleteSnapShot -eq $true ) {
+  #Deregister ami and delete snapshot Id
+  $ami = "$($env:BUILDIMAGE_AMIID)"
+  $snapshot = (Get-EC2Snapshot -owner self | Where-Object {$_.Description -like "*$ami*"}).SnapshotId
+  Write-Host "Deregistering the AMI"
+  Unregister-EC2Image -ImageId $ami -Force
+  Start-Sleep -Seconds 20
+  Write-Host "Deleting Snapshot"
+  Remove-EC2Snapshot -SnapshotId $snapshot -Force
+}
